Fix declaration stuff
