+++ tsconfig.json
@@ -64,6 +64,7 @@
   ],
   "include": [
     "next-env.d.ts",
+    "declaration.d.ts",
     "**/*.ts",
     "**/*.tsx",
     "**/*.js"
